Reorganize TransformMapBuilder
